Keep histories as int16
Elo   | 3.21 +- 2.10 (95%)
SPRT  | 2.0+0.02s Threads=1 Hash=8MB
LLR   | 2.90 (-2.25, 2.89) [0.00, 2.50]
Games | N: 28220 W: 7200 L: 6939 D: 14081
Penta | [165, 2968, 7604, 3187, 186]

Bench 3084378

+++ src/history.h
@@ -7,31 +7,31 @@ constexpr int CORRHIST_SIZE = 16384;
 constexpr int CORRHIST_LIMIT = 1024;
 
 // [color][from to]
-using MainHistory  = int[COLOR_NB][SQUARE_NB * SQUARE_NB];
+using MainHistory  = int16_t[COLOR_NB][SQUARE_NB * SQUARE_NB];
 
 // [piece to][piece_type]
-using CaptureHistory = int[PIECE_NB * SQUARE_NB][PIECE_TYPE_NB];
+using CaptureHistory = int16_t[PIECE_NB * SQUARE_NB][PIECE_TYPE_NB];
 
 // [piece to]
 using CounterMoveHistory = Move[PIECE_NB * SQUARE_NB];
 
 // [isCap][piece to][piece to]
-using ContinuationHistory = int[2][PIECE_NB * SQUARE_NB][PIECE_NB * SQUARE_NB];
+using ContinuationHistory = int16_t[2][PIECE_NB * SQUARE_NB][PIECE_NB * SQUARE_NB];
 
 // [stm][pawn hash]
-using PawnCorrHist = int[2][CORRHIST_SIZE];
+using PawnCorrHist = int16_t[2][CORRHIST_SIZE];
 
 // [stm][pawn hash]
-using NonPawnCorrHist = int[2][CORRHIST_SIZE];
+using NonPawnCorrHist = int16_t[2][CORRHIST_SIZE];
 
 inline int getCorrHistIndex(Key pawnKey){
   return pawnKey % CORRHIST_SIZE;
 }
 
-inline void addToCorrhist(int& history, int value){
-  history += value - history * abs(value) / CORRHIST_LIMIT;
+inline void addToCorrhist(int16_t& history, int value){
+  history += value - int(history) * abs(value) / CORRHIST_LIMIT;
 }
 
-inline void addToHistory(int& history, int value) {
-  history += value - history * abs(value) / 16384;
+inline void addToHistory(int16_t& history, int value) {
+  history += value - int(history) * abs(value) / 16384;
 }
